feat: Support `path` for `takeSnapshot` on Android
