+++ package/android/src/main/java/com/mrousavy/camera/core/utils/FileUtils.kt
@@ -4,11 +4,23 @@ import android.content.Context
 import android.graphics.Bitmap
 import android.graphics.BitmapFactory
 import android.util.Size
+import com.mrousavy.camera.core.InvalidPathError
 import java.io.File
 import java.io.FileOutputStream
 
 class FileUtils {
   companion object {
+    fun getDirectory(path: String?): File {
+      if (path == null) {
+        throw InvalidPathError("null")
+      }
+      val file = File(path)
+      if (!file.isDirectory) {
+        throw InvalidPathError(path)
+      }
+      return file
+    }
+
     fun createTempFile(context: Context, extension: String): File =
       File.createTempFile("mrousavy-", extension, context.cacheDir).also {
         it.deleteOnExit()
