+++ package/android/src/main/java/com/mrousavy/camera/react/CameraView+TakeSnapshot.kt
@@ -6,27 +6,29 @@ import com.facebook.react.bridge.WritableMap
 import com.mrousavy.camera.core.SnapshotFailedError
 import com.mrousavy.camera.core.SnapshotFailedPreviewNotEnabledError
 import com.mrousavy.camera.core.types.ShutterType
-import com.mrousavy.camera.core.types.SnapshotOptions
+import com.mrousavy.camera.core.types.TakeSnapshotOptions
 import com.mrousavy.camera.core.utils.FileUtils
 
 private const val TAG = "CameraView.takeSnapshot"
 
-fun CameraView.takeSnapshot(options: SnapshotOptions): WritableMap {
+fun CameraView.takeSnapshot(options: TakeSnapshotOptions): WritableMap {
   Log.i(TAG, "Capturing snapshot of Camera View...")
   val previewView = previewView ?: throw SnapshotFailedPreviewNotEnabledError()
   val bitmap = previewView.bitmap ?: throw SnapshotFailedError()
 
+  // Shutter Event (JS)
   onShutter(ShutterType.SNAPSHOT)
 
-  val file = FileUtils.createTempFile(context, ".jpg")
-  FileUtils.writeBitmapTofile(bitmap, file, options.quality)
+  // Write snapshot to .jpg file
+  FileUtils.writeBitmapTofile(bitmap, options.file, options.quality)
 
   Log.i(TAG, "Successfully saved snapshot to file!")
 
   val orientation = cameraSession.outputOrientation
 
+  // Parse output data
   val map = Arguments.createMap()
-  map.putString("path", file.absolutePath)
+  map.putString("path", options.file.absolutePath)
   map.putInt("width", bitmap.width)
   map.putInt("height", bitmap.height)
   map.putString("orientation", orientation.unionValue)
